optimize browser tests
