+++ packages/js-sdk/index.tests.js
@@ -2,6 +2,12 @@
 import {expect} from 'chai';
 import {assert} from 'chai';
 import testConfig from './tests/test-config.json';
+import {
+  getFeaturesWithStatuses,
+  getMultipleFeatureWithStatus,
+  getSingleFeatureWithStatus,
+  getVariationsAcrossAllExperiences
+} from './tests/setup/shared';
 
 class DataStore {
   data = {};
@@ -108,122 +114,46 @@ export default function runTests(bundle) {
       done();
     });
     it('Shoud successfully get variations across all experiences', function (done) {
-      const variationIds = ['100299456', '100299457', '100299460', '100299461'];
-      const variations = context.runExperiences({
-        locationProperties: {url: 'https://convert.com/'},
-        visitorProperties: {
-          varName3: 'something'
-        }
-      });
-      expect(variations).to.be.an('array').that.have.length(2);
-      variations.forEach((variation) =>
-        expect(variation)
-          .to.be.an('object')
-          .that.have.keys(
-            'experienceId',
-            'experienceKey',
-            'experienceName',
-            'bucketingAllocation',
-            'id',
-            'key',
-            'name',
-            'status',
-            'changes',
-            'is_baseline',
-            'traffic_allocation'
-          )
+      getVariationsAcrossAllExperiences(
+        {
+          accountId,
+          projectId,
+          context
+        },
+        done
       );
-      const selectedVariations = variations.map(({id}) => id);
-      expect(variationIds).to.include.deep.members(selectedVariations);
-      done();
     });
     it('Shoud successfully get a single feature and its status', function (done) {
-      const featureKey = 'feature-2';
       const featureId = '10025';
-      const feature = context.runFeature(featureKey, {
-        locationProperties: {url: 'https://convert.com/'},
-        visitorProperties: {
-          varName3: 'something'
-        }
-      });
-      expect(feature)
-        .to.be.an('object')
-        .that.have.keys(
-          'experienceId',
-          'experienceKey',
-          'experienceName',
-          'id',
-          'key',
-          'name',
-          'status',
-          'variables'
-        );
-      expect(feature.id).to.equal(featureId);
-      done();
+      getSingleFeatureWithStatus(
+        {
+          accountId,
+          projectId,
+          featureId,
+          context
+        },
+        done
+      );
     });
     it('Shoud successfully get multiple features and its status', function (done) {
-      const featureKey = 'feature-1';
-      const featureIds = ['10024', '10025'];
-      const features = context.runFeature(featureKey, {
-        locationProperties: {url: 'https://convert.com/'},
-        visitorProperties: {
-          varName3: 'something'
-        }
-      });
-      expect(features).to.be.an('array').that.have.length(2);
-      features.forEach((feature) =>
-        expect(feature)
-          .to.be.an('object')
-          .that.have.keys(
-            'experienceId',
-            'experienceKey',
-            'experienceName',
-            'id',
-            'key',
-            'name',
-            'status',
-            'variables'
-          )
+      getMultipleFeatureWithStatus(
+        {
+          accountId,
+          projectId,
+          context
+        },
+        done
       );
-      const selectedFeatures = features.map(({id}) => id);
-      expect(featureIds).to.include.deep.members(selectedFeatures);
-      done();
     });
     it('Shoud successfully get features and their statuses', function (done) {
-      const featureIds = ['10024', '10025', '10026'];
-      const features = context.runFeatures({
-        locationProperties: {url: 'https://convert.com/'},
-        visitorProperties: {
-          varName3: 'something'
-        }
-      });
-      expect(features).to.be.an('array').that.have.length(4);
-      features
-        .filter(({status}) => status === 'enabled')
-        .forEach((feature) =>
-          expect(feature)
-            .to.be.an('object')
-            .that.have.keys(
-              'experienceId',
-              'experienceKey',
-              'experienceName',
-              'id',
-              'key',
-              'name',
-              'status',
-              'variables'
-            )
-        );
-      features
-        .filter(({status}) => status === 'disabled')
-        .forEach((feature) =>
-          expect(feature)
-            .to.be.an('object')
-            .that.have.keys('id', 'key', 'name', 'status')
-        );
-      const selectedFeatures = features.map(({id}) => id);
-      expect(featureIds).to.include.deep.members(selectedFeatures);
-      done();
+      getFeaturesWithStatuses(
+        {
+          accountId,
+          projectId,
+          context
+        },
+        done
+      );
     });
     it('Should trigger Conversion', function () {
       const goalKey = 'increase-engagement';
