+++ packages/js-sdk/tests/context.tests.ts
@@ -18,6 +18,12 @@ import {Config as ConfigType} from '@convertcom/js-sdk-types';
 import {objectDeepMerge} from '@convertcom/js-sdk-utils';
 import {EntityType} from '@convertcom/js-sdk-enums';
 import {defaultConfig} from '../src/config/default';
+import {
+  getFeaturesWithStatuses,
+  getMultipleFeatureWithStatus,
+  getSingleFeatureWithStatus,
+  getVariationsAcrossAllExperiences
+} from './setup/shared';
 
 const host = 'http://localhost';
 const port = 8090;
@@ -163,156 +169,52 @@ describe('Context tests', function () {
     });
     it('Shoud successfully get variations across all experiences', function (done) {
       this.timeout(test_timeout);
-      const variationIds = ['100299456', '100299457', '100299460', '100299461'];
-      const variations = context.runExperiences({
-        locationProperties: {url: 'https://convert.com/'},
-        visitorProperties: {
-          varName3: 'something'
-        }
-      });
-      server.on('request', (request, res) => {
-        if (request.url.startsWith(`/track/${accountId}/${projectId}`)) {
-          request.on('end', () => {
-            expect(variations).to.be.an('array').that.have.length(2);
-            variations.forEach((variation) =>
-              expect(variation)
-                .to.be.an('object')
-                .that.have.keys(
-                  'experienceId',
-                  'experienceKey',
-                  'experienceName',
-                  'bucketingAllocation',
-                  'id',
-                  'key',
-                  'name',
-                  'status',
-                  'changes',
-                  'is_baseline',
-                  'traffic_allocation'
-                )
-            );
-            const selectedVariations = variations.map(({id}) => id);
-            expect(variationIds).to.include.deep.members(selectedVariations);
-            done();
-          });
-        }
-        res.writeHead(200, {'Content-Type': 'application/json'});
-        res.end('{}');
-      });
+      getVariationsAcrossAllExperiences(
+        {
+          accountId,
+          projectId,
+          context,
+          server
+        },
+        done
+      );
     });
     it('Shoud successfully get a single feature and its status', function (done) {
       this.timeout(test_timeout);
-      const featureKey = 'feature-2';
-      const feature = context.runFeature(featureKey, {
-        locationProperties: {url: 'https://convert.com/'},
-        visitorProperties: {
-          varName3: 'something'
-        }
-      });
-      server.on('request', (request, res) => {
-        if (request.url.startsWith(`/track/${accountId}/${projectId}`)) {
-          request.on('end', () => {
-            expect(feature)
-              .to.be.an('object')
-              .that.have.keys(
-                'experienceId',
-                'experienceKey',
-                'experienceName',
-                'id',
-                'key',
-                'name',
-                'status',
-                'variables'
-              );
-            expect(feature.id).to.equal(featureId);
-            done();
-          });
-        }
-        res.writeHead(200, {'Content-Type': 'application/json'});
-        res.end('{}');
-      });
+      getSingleFeatureWithStatus(
+        {
+          accountId,
+          projectId,
+          featureId,
+          context,
+          server
+        },
+        done
+      );
     });
     it('Shoud successfully get multiple features and its status', function (done) {
       this.timeout(test_timeout);
-      const featureKey = 'feature-1';
-      const featureIds = ['10024', '10025'];
-      const features = context.runFeature(featureKey, {
-        locationProperties: {url: 'https://convert.com/'},
-        visitorProperties: {
-          varName3: 'something'
-        }
-      });
-      server.on('request', (request, res) => {
-        if (request.url.startsWith(`/track/${accountId}/${projectId}`)) {
-          request.on('end', () => {
-            expect(features).to.be.an('array').that.have.length(2);
-            features.forEach((feature) =>
-              expect(feature)
-                .to.be.an('object')
-                .that.have.keys(
-                  'experienceId',
-                  'experienceKey',
-                  'experienceName',
-                  'id',
-                  'key',
-                  'name',
-                  'status',
-                  'variables'
-                )
-            );
-            const selectedFeatures = features.map(({id}) => id);
-            expect(featureIds).to.include.deep.members(selectedFeatures);
-            done();
-          });
-        }
-        res.writeHead(200, {'Content-Type': 'application/json'});
-        res.end('{}');
-      });
+      getMultipleFeatureWithStatus(
+        {
+          accountId,
+          projectId,
+          context,
+          server
+        },
+        done
+      );
     });
     it('Shoud successfully get features and their statuses', function (done) {
       this.timeout(test_timeout);
-      const featureIds = ['10024', '10025', '10026'];
-      const features = context.runFeatures({
-        locationProperties: {url: 'https://convert.com/'},
-        visitorProperties: {
-          varName3: 'something'
-        }
-      });
-      server.on('request', (request, res) => {
-        if (request.url.startsWith(`/track/${accountId}/${projectId}`)) {
-          request.on('end', () => {
-            expect(features).to.be.an('array').that.have.length(4);
-            features
-              .filter(({status}) => status === 'enabled')
-              .forEach((feature) =>
-                expect(feature)
-                  .to.be.an('object')
-                  .that.have.keys(
-                    'experienceId',
-                    'experienceKey',
-                    'experienceName',
-                    'id',
-                    'key',
-                    'name',
-                    'status',
-                    'variables'
-                  )
-              );
-            features
-              .filter(({status}) => status === 'disabled')
-              .forEach((feature) =>
-                expect(feature)
-                  .to.be.an('object')
-                  .that.have.keys('id', 'key', 'name', 'status')
-              );
-            const selectedFeatures = features.map(({id}) => id);
-            expect(featureIds).to.include.deep.members(selectedFeatures);
-            done();
-          });
-        }
-        res.writeHead(200, {'Content-Type': 'application/json'});
-        res.end('{}');
-      });
+      getFeaturesWithStatuses(
+        {
+          accountId,
+          projectId,
+          context,
+          server
+        },
+        done
+      );
     });
     it('Should trigger Conversion', function (done) {
       this.timeout(test_timeout);
